+++ src/Models/Cars/errors.js
@@ -1,4 +1,4 @@
-import RuntimeError from '../../RuntimeError';
+import RuntimeError from '../../RuntimeError.js';
 
 export class DuplicatedCarNameError extends RuntimeError {
     static MESSAGE = '중복된 자동차 이름으로는 프로그램이 동작할 수 없습니다.';
