+++ src/Models/MoveStrategy/RandomStrategy/errors.js
@@ -1,0 +1,17 @@
+import RuntimeError from '../../../RuntimeError.js';
+
+export class MinMaxNumberIsNotNumberError extends RuntimeError {
+    static MESSAGE = 'min, max는 숫자여야 합니다.';
+
+    constructor() {
+        super(MinMaxNumberIsNotNumberError.MESSAGE);
+    }
+}
+
+export class MinNumberGreaterThanMaxNumberError extends RuntimeError {
+    static MESSAGE = 'min은 max 보다 클 수 없습니다.';
+
+    constructor() {
+        super(MinNumberGreaterThanMaxNumberError.MESSAGE);
+    }
+}
