+++ test/FixedStrategy.js
@@ -1,5 +1,5 @@
-import MoveStrategy from '../src/Models/MoveStrategy';
-import { FixedNumberIsNotNumberError } from './errors';
+import MoveStrategy from '../src/Models/MoveStrategy/MoveStrategy.js';
+import { FixedNumberIsNotNumberError } from './errors.js';
 
 export default class FixedStrategy extends MoveStrategy {
     #number;
